refactor: minor fixes and refactors
