+++ README.md
@@ -1,5 +1,5 @@
 # Phaseball
 
- [![coverage report](https://hq.0xa1.red/axdx/phaseball/badges/trunk/coverage.svg)](https://hq.0xa1.red/axdx/phaseball/-/commits/trunk) 
+ [![coverage report](https://github.com/0xa1-red/phaseball/badges/trunk/coverage.svg)](https://github.com/0xa1-red/phaseball/-/commits/trunk) 
 
 Baseball story generator based on W. M. Akers' Deadball
