+++ _examples/team-gen/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 
-	"hq.0xa1.red/axdx/phaseball/internal/dice"
+	"github.com/0xa1-red/phaseball/internal/dice"
 )
 
 func main() {
